backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sse-starlette>=1.8.0
pydantic>=2.0
python-dotenv>=1.0.0
